feat: utili % and refactor
